subida codico v3 terminado a falta de comprobaciones

+++ encriptar.py
@@ -5,6 +5,8 @@ import tkinter as tk
 import os as os
 import sys
 import hashlib
+from cryptography.fernet import Fernet
+
 
 def limpiar_padding(lectura, escritura):
     with open(lectura, 'r') as infile, open(escritura, 'w') as outfile:
@@ -52,43 +54,64 @@ def encrypt_file(input_file, key):
     ciphertext_block = encryptor.finalize()
     os.remove(input_file)
     
-def decrypt_file(input_file, key):
-    if ".cif" not in input_file:
-        print("No se puede desencriptar un archivo no cifrado")
-        return
+def decrypt_file(input_file, key, funcionamiento):
+    if funcionamiento == 1:
+        if ".cif" not in input_file:
+            print("No se puede desencriptar un archivo no cifrado")
+            return
      
-    output_file=os.path.splitext(input_file)[0]
+        output_file=os.path.splitext(input_file)[0]
     # Tamaño del bloque AES en bytes (128 bits)
     block_size = 16
     
     # Crea un objeto AES Cipher con la clave proporcionada y modo de operación CBC
     cipher = Cipher(algorithms.AES(key), modes.CBC(b'\0' * block_size), backend=default_backend())
     decryptor = cipher.decryptor()
+    clave = b''
+    if funcionamiento == 1:
+        # Abre el archivo cifrado y el archivo de salida en modo binario
+        with open(input_file, 'rb') as infile, open(output_file, 'wb') as outfile:
+            # Lee y descifra el archivo en bloques
+            while True:
+                # Lee un bloque cifrado del archivo de entrada
+                ciphertext_block = infile.read(block_size)
+                if not ciphertext_block:
+                    break  # Fin del archivo
+                
+                # Descifra el bloque y escribe el resultado en el archivo de salida
+                plaintext_block = decryptor.update(ciphertext_block)
+                
+                outfile.write(plaintext_block)
+                
+    else:
+        
+        for i in range(0, len(input_file), block_size):
 
-    # Abre el archivo cifrado y el archivo de salida en modo binario
-    with open(input_file, 'rb') as infile, open(output_file, 'wb') as outfile:
-        # Lee y descifra el archivo en bloques
-        while True:
-            # Lee un bloque cifrado del archivo de entrada
-            ciphertext_block = infile.read(block_size)
+            # Obtiene un bloque cifrado de los datos de entrada
+            ciphertext_block = input_file[i:i+block_size]
             if not ciphertext_block:
-                break  # Fin del archivo
-            
-            # Descifra el bloque y escribe el resultado en el archivo de salida
+                break  # Fin de los datos
+
+            # Descifra el bloque
             plaintext_block = decryptor.update(ciphertext_block)
-            outfile.write(plaintext_block)
-    
+
+            # Añade el bloque descifrado a la clave
+            clave += plaintext_block
+                
     # Finaliza el descifrado (puede haber datos adicionales en el buffer del descifrador)
     plaintext_block = decryptor.finalize()
 
+    if funcionamiento == 1:
+        # Abre el archivo de salida en modo escritura y recorta el relleno
+        limpiar_padding(output_file, "final")
 
-   # Abre el archivo de salida en modo escritura y recorta el relleno
-    limpiar_padding(output_file, "final")
-
-    temporal = output_file
-    os.remove(output_file)
-    os.rename("final", temporal)
-    os.remove(input_file)
+    
+        temporal = output_file
+        os.remove(output_file)
+        os.rename("final", temporal)
+        os.remove(input_file)
+    else:    
+        return clave
 
 def hash_texto(texto_plano):
     # Crear un objeto hash usando SHA256
@@ -110,6 +133,18 @@ def verificar_contraseña(contraseña):
     else:
         return False
 
+
+
+
+def obtener_primeros_256_bits(cadena_bytes):
+    # Asegurarse de que la cadena de bytes tiene al menos 32 bytes (256 bits)
+    assert len(cadena_bytes) >= 32, "La cadena de bytes es demasiado corta"
+
+    # Obtener los primeros 32 bytes (256 bits)
+    primeros_256_bits=b''
+    primeros_256_bits = cadena_bytes[:32]
+
+    return primeros_256_bits
 
 #if __name__ == "__main__":
    # key = b'\x01\x23\x45\x67\x89\xab\xcd\xef\xfe\xdc\xba\x98\x76\x54\x32\x10' \
